Module type Basic_intf.Reals

Module type of anything that looks like the (measured) reals.

include Field_std

include Field

include Ring

include Abelian_group

type t

val zero : t

val add : t -> t -> t

val neg : t -> t

include Monoid with type t := t

type t

val one : t

val mul : t -> t -> t

val sub : t -> t -> t

val of_int : int -> t

val div : t -> t -> t

include Basic_structures__.Intf_std.Std with type t := t

type t

val compare : t -> t -> int

val equal : t -> t -> bool

val pp : t Basic_structures__.Intf_std.printer

val hash : t -> int

val lebesgue : Stdlib.Random.State.t -> t

Lebesgue on the interval 0;1

val (+) : t -> t -> t

val (-) : t -> t -> t

val (*) : t -> t -> t

val (/) : t -> t -> t

include Infix_order with type t := t and type 'a m := 'a

type 'a m

type t

val (<) : t m -> t m -> bool m

val (>) : t m -> t m -> bool m

val (<=) : t m -> t m -> bool m

val (>=) : t m -> t m -> bool m

val (=) : t m -> t m -> bool m

val (<>) : t m -> t m -> bool m

val npow : t -> int -> t

npow x n is x tp the power of n.

val to_float : t -> float

val of_float : float -> t

val of_int : int -> t