Module type Basic_intf.Metric

Metric space

include Std

type t

val compare : t -> t -> int

val equal : t -> t -> bool

val pp : t printer

val hash : t -> int

val dist : t -> t -> float

dist is expected to be a proper distance function (symmetric, zero on the diagonal, verifying the triangular inequality).