Module Basic_structures__Free_module

Free module over a ring generated by a finite, totally ordered set

module Make : functor (Basis : Basic_structures.Basic_intf.Std) -> functor (R : Basic_structures.Basic_intf.Ring_std) -> functor (M : Stdlib.Map.S with type Make.key = Basis.t) -> Basic_structures.Basic_intf.Free_module_std with module R = R and module Basis = Basis and type basis = Basis.t and type t = R.t M.t

Make constructs an (OCaml) module to handle the free (algebraic) R-module generated by finite subsets of Basis. The implementation of free modules is backed by a map implementation M that must be provided by the user.

module type Free_module_with_map = sig ... end

A free module packed with its underlying map implementation

module Make_with_map : functor (X : Basic_structures.Basic_intf.Std) -> functor (R : Basic_structures.Basic_intf.Ring_std) -> Free_module_with_map with type r = R.t and type Map.key = X.t

Make_with_map constructs a free R-module over X, instsan

module Rational_valued : sig ... end

Free modules with coefficients in Q.t

module Float_valued : sig ... end

Free modules with float coefficients