Module type Basic_structures__Intf_monad.Codegen_monad

Monad dedicated to handling computations spanning several stages

type 'a m

'a m is the type of generated programs of type 'a

type 'a t

'a t is the type of computations of type 'a

val return : 'a -> 'a t

val bind : 'a t -> ('a -> 'b t) -> 'b t

val run : 'a m t -> 'a m

val lift1 : ('a -> 'b) -> 'a t -> 'b t

val lift2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val lift3 : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

module Infix : sig ... end